refactor(dialogue_manager): extract background.rs
Step 5/9. Move the spawn-and-await machinery into a dedicated module:
- await_pending_classification (test-only): drain classifier alone
- await_pending_background: drain both tasks concurrently via join!
- drain_classification / drain_post_response: take JoinHandle out of
  self and return a 'static future awaiting it with timeout
- apply_classification_outcome: apply EngagementAssessment with
  monotonic-dirty gating
- apply_post_response_outcome: apply extraction (insert) → comprehension
  (promote-via-max) and merge concepts into in-memory turns

All entry points are pub(super); behavior unchanged.

mod.rs: 2773 → 2624 lines.
background.rs: new, 173 lines.
